+++ src/controllers/FoodDataController.ts
@@ -6,17 +6,29 @@ import {
   updateFoodDataById,
   deleteFoodDataById,
   getFoodDataBySearch,
+  generateFoodStats,
 } from '../models/FoodDataModel';
 import { parseDatabaseError } from '../utils/db-utils';
 import { UserIdParam } from '../types/userInfo';
 import { FoodData as FoodEntity } from '../entities/foodData';
+import { getUserById } from '../models/UserModel';
 
 async function submitFoodData(req: Request, res: Response): Promise<void> {
   const foodData = req.body as FoodData;
 
+  // check that user is logged in
+  if (!req.session.isLoggedIn) {
+    res.sendStatus(401);
+    return;
+  }
+
+  // use session to add data for that user only
+  const user = await getUserById(req.session.authenticatedUser.userId);
   try {
-    await addFoodData(foodData);
-    res.sendStatus(201);
+    if (user) {
+      const food = await addFoodData(foodData, user);
+      res.status(201).json(food);
+    }
   } catch (err) {
     console.error(err);
     const databaseErrorMessage = parseDatabaseError(err);
@@ -27,11 +39,20 @@ async function submitFoodData(req: Request, res: Response): Promise<void> {
 async function getAllUserFoodData(req: Request, res: Response): Promise<void> {
   const { userId } = req.params as unknown as UserIdParam;
 
-  // include check for userid validity here once implemented
+  // check that user is logged in
+  if (!req.session.isLoggedIn) {
+    res.sendStatus(401);
+    return;
+  }
+
+  // check that user accesses only their data based on session
+  if (!(userId === req.session.authenticatedUser.userId)) {
+    res.sendStatus(403);
+    return;
+  }
 
   try {
     const foodData = await getAllFoodDataForUser(userId);
-    console.log(foodData);
     res.json(foodData);
   } catch (err) {
     console.error(err);
@@ -43,6 +64,12 @@ async function getAllUserFoodData(req: Request, res: Response): Promise<void> {
 async function getFoodData(req: Request, res: Response): Promise<void> {
   const { foodDataId } = req.params as unknown as FoodDataIdParam;
 
+  // check that user is logged in
+  if (!req.session.isLoggedIn) {
+    res.sendStatus(401);
+    return;
+  }
+
   try {
     const foodData = await getFoodDataById(foodDataId);
     if (!foodData) {
@@ -50,7 +77,11 @@ async function getFoodData(req: Request, res: Response): Promise<void> {
       res.sendStatus(404);
       return;
     }
-    console.log(foodData);
+    // check that this fooddata item belongs to this user
+    if (!(foodData?.user.userId === req.session.authenticatedUser.userId)) {
+      res.sendStatus(403);
+      return;
+    }
     res.json(foodData);
   } catch (err) {
     console.error(err);
@@ -63,14 +94,26 @@ async function updateFoodData(req: Request, res: Response): Promise<void> {
   const { foodDataId } = req.params as unknown as FoodDataIdParam;
   const newFood = req.body as FoodData;
 
+  if (!req.session.isLoggedIn) {
+    // check that user is logged in
+    res.sendStatus(401);
+    return;
+  }
+
   try {
-    const foodData = await updateFoodDataById(foodDataId, newFood);
-    if (!foodData) {
+    const existingData = await getFoodDataById(foodDataId);
+    if (!existingData) {
       // not found
       res.sendStatus(404);
       return;
     }
-    console.log(foodData);
+    // check that this fooddata item belongs to this user
+    if (!(existingData?.user.userId === req.session.authenticatedUser.userId)) {
+      res.sendStatus(403);
+      return;
+    }
+
+    const foodData = await updateFoodDataById(foodDataId, newFood);
     res.json(foodData);
   } catch (err) {
     console.error(err);
@@ -82,7 +125,25 @@ async function updateFoodData(req: Request, res: Response): Promise<void> {
 async function deleteFoodData(req: Request, res: Response): Promise<void> {
   const { foodDataId } = req.params as unknown as FoodDataIdParam;
 
+  if (!req.session.isLoggedIn) {
+    // check that user is logged in
+    res.sendStatus(401);
+    return;
+  }
+
   try {
+    const foodData = await getFoodDataById(foodDataId);
+    if (!foodData) {
+      // not found
+      res.sendStatus(404);
+      return;
+    }
+    // check that this fooddata item belongs to this user
+    if (!(foodData?.user.userId === req.session.authenticatedUser.userId)) {
+      res.sendStatus(403);
+      return;
+    }
+
     await deleteFoodDataById(foodDataId);
     res.sendStatus(200);
   } catch (err) {
@@ -95,14 +156,44 @@ async function deleteFoodData(req: Request, res: Response): Promise<void> {
 async function searchFoodData(req: Request, res: Response): Promise<void> {
   const { start, end, keyword } = req.query as FoodSearchParam;
 
+  // check that user is logged in
+  if (!req.session.isLoggedIn) {
+    res.sendStatus(401);
+    return;
+  }
   if (start && end && start > end) {
     res.sendStatus(400); // invalid start/end times
     return;
   }
 
   try {
-    const foodData: FoodEntity[] = await getFoodDataBySearch(start, end, keyword);
+    const foodData: FoodEntity[] = await getFoodDataBySearch(
+      req.session.authenticatedUser.userId,
+      start,
+      end,
+      keyword,
+    );
     res.json(foodData);
+  } catch (err) {
+    console.error(err);
+    const databaseErrorMessage = parseDatabaseError(err);
+    res.status(500).json(databaseErrorMessage);
+  }
+}
+
+async function getFoodStats(req: Request, res: Response): Promise<void> {
+  const { start, end } = req.query as FoodSearchParam;
+  if (!start || !end || start > end) {
+    res.sendStatus(400); // invalid start/end times
+    return;
+  }
+  try {
+    const stats: FoodStats[] = await generateFoodStats(
+      req.session.authenticatedUser.userId,
+      start,
+      end,
+    );
+    res.json(stats);
   } catch (err) {
     console.error(err);
     const databaseErrorMessage = parseDatabaseError(err);
@@ -117,4 +208,5 @@ export default {
   updateFoodData,
   deleteFoodData,
   searchFoodData,
+  getFoodStats,
 };
